requirements.txt
streamlit>=1.28.0
openai>=1.40.0
anthropic>=0.34.0
python-dotenv>=1.0.0
pypdf>=4.0.0
python-docx>=1.1.0
